+++ packages/wasmio/test/basic.spec.ts
@@ -1,4 +1,4 @@
-import echo from 'pkg'
+import echo from 'wasmio'
 
 test('it should just work', () => {
   expect(echo()).toBe('Hello World!')