+++ vitest.config.ts
@@ -1,0 +1,11 @@
+import autoImport from 'unplugin-auto-import/vite'
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+  plugins: [
+    autoImport({
+      imports: ['vitest'],
+      dts: true,
+    }),
+  ],
+})