+++ CHANGELOG.md
@@ -3,4 +3,6 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-- [pkg](./packages/pkg/CHANGELOG.md)
+- [@wasmio/go](./packages/@wasmio/go/CHANGELOG.md)
+- [@wasmio/tinygo](./packages/@wasmio/tinygo/CHANGELOG.md)
+- [wasmio](./packages/wasmio/CHANGELOG.md)